[refactor] : 코드리뷰 반영 및 리팩토링
