--- survey/application/src/main/java/me/nalab/survey/application/port/in/web/SurveyFindUseCase.java
@@ -1,8 +1,0 @@
-package me.nalab.survey.application.port.in.web;
-
-import me.nalab.survey.application.common.dto.SurveyDto;
-
-public interface SurveyFindUseCase {
-	SurveyDto findSurvey(Long surveyId);
-
-}
