--- survey/application/src/main/java/me/nalab/survey/application/port/in/web/TargetFindUseCase.java
@@ -1,7 +1,0 @@
-package me.nalab.survey.application.port.in.web;
-
-import me.nalab.survey.application.common.dto.TargetDto;
-
-public interface TargetFindUseCase {
-	TargetDto findTarget(Long targetId);
-}
