+++ survey/application/src/main/java/module-info.java
@@ -3,6 +3,8 @@ module luffy.survey.application.main {
 	exports me.nalab.survey.application.common.dto;
 	exports me.nalab.survey.application.port.in.web;
 	exports me.nalab.survey.application.port.out.persistence;
+	exports me.nalab.survey.application.port.in.web.survey.find;
+	exports me.nalab.survey.application.port.in.web.target.find;
 
 	requires luffy.survey.domain.main;
 	requires luffy.core.id.generator.id.generator.starter.main;
